feat: imple logic
