+++ crates/rome_js_analyze/src/lib.rs
@@ -15,6 +15,7 @@ use std::{borrow::Cow, error::Error};
 mod analyzers;
 mod aria_analyzers;
 mod aria_services;
+mod aria_utils;
 mod assists;
 mod ast_utils;
 mod control_flow;
